refactor start/project code
