+++ packages/expo/cli/start/project/settings.ts
